requirements.txt
av>=9.0.0
imagehash>=4.2.1
loguru>=0.5.3
m3u8>=3.0.0
mutagen>=1.46.0
pillow>=8.4.0
plyvel-ci>=1.5.0
psutil>=5.9.0
pyexiv2>=2.8.2
python-dateutil>=2.8.2
requests>=2.26.0
rich>=13.0.0
mutagen>=1.46.0
pyexiv2>=2.8.2
